requirements.txt
asgiref==3.7.2
b2sdk==1.30.1
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.9
django-appconf==1.0.6
django-autoslug==1.9.9
django-cities-light==3.10.1
django-countries==7.5.1
django-cropper-image==1.0.5
django-image-cropping==1.7
django-phonenumber-field==7.3.0
easy-thumbnails==2.8.5
idna==3.6
logfury==1.0.1
packaging==23.2
phonenumberslite==8.13.28
pillow==10.2.0
progressbar2==4.3.2
psycopg==3.1.17
psycopg-binary==3.1.17
python-http-client==3.3.7
python-magic==0.4.27
python-utils==3.8.2
pytz==2024.1
requests==2.31.0
sendgrid==6.11.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tqdm==4.66.2
typing_extensions==4.9.0
Unidecode==1.3.8
urllib3==2.2.1
